requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
pydantic
typing-extensions
numpy
requests
aiofiles
python-multipart
aiohttp
motor
pymongo
openai